requirements.txt
requests
google-generativeai
keyboard
python-dotenv